+++ documentation/changelog.markdown
@@ -6,6 +6,7 @@ To be determined.
 
 ## Details
 - **New**: Added support for styled text in all titles, labels, and text layers.
+- **New**: Added a minor tick label shadow property to `CPTAxis`.
 - **Changed**: Miscellaneous bug fixes and cleanup.